apps/server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aequitas-server"
version = "0.1.0"
description = "Aequitas FI API — FastAPI + LangGraph orchestration"
requires-python = ">=3.11"
dependencies = [
  "aequitas-database",
  "aequitas-ai",
  "anthropic>=0.30.0",
  "presidio-analyzer>=2.2.0",
  "spacy>=3.7.0",
  "apscheduler>=3.10.0",
  "fastapi[standard]>=0.115.0",
  "llama-parse>=0.4.0",
  "openai>=1.40.0",
  "pymupdf>=1.24.0",
  "unstructured[pdf]>=0.10.0",
  "uvicorn[standard]>=0.32.0",
  "sqlalchemy>=2.0.36",
  "asyncpg>=0.30.0",
  "langgraph>=0.2.0",
  "langgraph-checkpoint-postgres>=2.0.0",
  "psycopg[binary]>=3.2.0",
  "langchain-core>=0.3.0",
  "langchain-openai>=0.2.0",
  "langchain-anthropic>=0.2.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.6.0",
  "httpx>=0.28.0",
  "python-multipart>=0.0.9",
  "nest-asyncio>=1.6.0",
  "pyjwt[crypto]>=2.8.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.8.0",
  "pytest>=8.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app", "api", "middleware"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
